% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours_caf}
\alias{colours_caf}
\title{Colour code used for the INTER grant application to plot the expression of genes of interest in epithelial vs CAF cells.}
\usage{
colours_caf(epithelium = "epithelium", caf = "caf")
}
\description{
Colour code used for the INTER grant application to plot the expression of genes of interest in epithelial vs CAF cells.
}
