% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours_cms}
\alias{colours_cms}
\title{The colours associated to the the consensus molecular subtypes (CMS) as used in the original publication.}
\usage{
colours_cms()
}
\description{
The colours associated to the the consensus molecular subtypes (CMS) as used in the original publication.
}
