% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_expr_score.R
\name{get_top_expressor}
\alias{get_top_expressor}
\title{Identify top-expressors based on the top-expressor score}
\usage{
get_top_expressor(z_scores, threshold = 0.5)
}
\arguments{
\item{z_scores}{vector of numericals containing the z-scores of gene signatures.}

\item{threshold}{numerical telling the proportion of signature genes that should be highly expressed}
}
\description{
(See the CAF signature repository for more details)
}
