% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_score.R
\name{ks_score}
\alias{ks_score}
\title{Kolmogorov-Smirnov test score}
\usage{
ks_score(x, y, na.rm = FALSE)
}
\arguments{
\item{x, y}{vectors of numericals containing the gene expressions to compare.}
}
\description{
Resume a multiple gene expression difference as a Kolmogorov-Smirnov Test based score
}
