% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_date.R
\name{last_updated}
\alias{last_updated}
\title{Cluster a tibble}
\usage{
last_updated(created = NULL, rmd = NULL)
}
\arguments{
\item{created}{NULL or character which will be parsed as the creation date}
}
\description{
Coerce the specified variables to a matrix, perform the clustering and return the tibble rearranged
according to the clustered ordering.
}
