% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnm_stage.R
\name{tnm_to_stage}
\alias{tnm_to_stage}
\title{Cancer stage grouping}
\usage{
tnm_to_stage(t, n, m, substage = TRUE, na_as_x = FALSE, mx_as_m0 = FALSE)
}
\arguments{
\item{t, n, m}{TNM (tumor - node - metastasis) staging system}

\item{strict}{If FALSE returns also approximate staging for missing n or m values.
nx or mx will be considered as not detected (n0 and m0) and the stage will be at least the returned value.}

\item{.prefix, .suffix}{The prefix and/or suffix used if strict is FALSE and n or m are not assessable (nx or mx). Defaults to \code{.suffix="+"} (stage is at least the calculated value)}
}
\description{
Group the T, N, and M classifications as clinical stages. The used corresponance is provided on page 3 of the \emph{outcomes data form} report (\emph{final version 2.0, 14/06/2012}).
}
\details{
t, n and m stage grouping as described in the \emph{outcomes data form} report:
\tabular{rrrr}{
  stage \tab t \tab n \tab m \cr
  I \tab T1, T2 \tab N0 \tab M0\cr
  II \tab T3, T4 \tab N0 \tab M0\cr
  II \tab A T3 \tab N0 \tab M0\cr
  II \tab B T4a \tab N0 \tab M0\cr
  II \tab C T4b \tab N0 \tab M0\cr
  III \tab Any T \tab N1, N2 \tab M0\cr
  III \tab A T1, T2 \tab N1 \tab M0\cr
  III \tab A T1 \tab N2a \tab M0\cr
  III \tab B T3, T4a \tab N1 \tab M0\cr
  III \tab B T2, T3 \tab N2a \tab M0\cr
  III \tab B T1, T2 \tab N2b \tab M0\cr
  III \tab C T4a \tab N2a \tab M0\cr
  III \tab C T3, T4a \tab N2b \tab M0\cr
  III \tab C T4b \tab N1, N2 \tab M0\cr
  IV A \tab Any T \tab Any N \tab M1a\cr
  IV B \tab Any T \tab Any N \tab M1b
}
}
\examples{

tnm_to_stage("t1", "n0", "m0")
tnm_to_stage("t1", "n0", "mx")
tnm_to_stage("t1", "n0", "mx", strict = FALSE)

}
