% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_expr_score.R
\name{top_expressor_score}
\alias{top_expressor_score}
\title{Top-expressor score}
\usage{
top_expressor_score(z_scores, scale = FALSE)
}
\arguments{
\item{z_scores}{vector of numericals containing the z-scores of gene signatures.}

\item{scale}{logical telling if the score should be returned as a raw value or scaled}
}
\description{
(See the CAF signature repository for more details)
}
